requirements.txt
selenium
Pillow
requests
numpy